Basic support for ray tracing with Metal on macOS
Update externals build script to build Slang locally

Fix Mac Build (#180)

Simplify Aurora MacOS Build (#181)

Script fixes

Code fixes

Doc updates

Fix clang regression

Clean up script

Support interactive display in Plasma on macOS (#184)

Restores Slang shader files and fixes HgiRenderer (#187)

Vulkan backend fixes (#188)

Resize window within macbook's screen

Fix Windows tests

Switch to public USD

Add sample model

+++ Libraries/Aurora/Source/MaterialBase.h
@@ -31,7 +31,7 @@ public:
 
     TextureProperties(const vector<TextureIdentifier>& names)
     {
-        for (int i = 0; i < names.size(); i++)
+        for (size_t i = 0; i < names.size(); i++)
         {
             auto& name                     = names[i];
             _textureNameLookup[name.image] = int(_properties.size());
