+++ Libraries/Aurora/Source/MaterialDefinition.cpp
@@ -19,7 +19,7 @@ MaterialDefaultValues::MaterialDefaultValues(const UniformBufferDefinition& prop
 {
     AU_ASSERT(
         defaultProps.size() == propertyDefs.size(), "Default properties do not match definition");
-    for (int i = 0; i < defaultTxt.size(); i++)
+    for (size_t i = 0; i < defaultTxt.size(); i++)
     {
         textureNames.push_back(defaultTxt[i].name);
     }
